third_party/blink/web_tests/fast/forms/button-type-change-usecounter.html

<!DOCTYPE html>
<link rel="author" href="mailto:ansollan@microsoft.com">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="container"></div>

<script>
const kTypeChangedConnected = 5815;
const kTypeChangedDisconnected = 5817;

function resetUseCounters() {
  internals.clearUseCounter(document, kTypeChangedConnected);
  internals.clearUseCounter(document, kTypeChangedDisconnected);
}

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  button.setAttribute('type', 'button');

  assert_false(internals.isUseCounted(document, kTypeChangedConnected));
  assert_true(internals.isUseCounted(document, kTypeChangedDisconnected));
}, 'Type set via setAttribute before connection only triggers disconnected UseCounter');

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  container.appendChild(button);
  button.setAttribute('type', 'button');

  assert_true(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
  button.remove();
}, 'Type set via setAttribute after connection only triggers connected UseCounter');

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  button.type = 'button';

  assert_false(internals.isUseCounted(document, kTypeChangedConnected));
  assert_true(internals.isUseCounted(document, kTypeChangedDisconnected));
}, 'Type set via property before connection only triggers disconnected UseCounter');

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  container.appendChild(button);
  button.type = 'button';

  assert_true(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
  button.remove();
}, 'Type set via property after connection only triggers connected UseCounter');

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  button.setAttribute('type', 'submit');

  assert_false(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
}, 'Setting same type via setAttribute while disconnected does not trigger UseCounters');

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  container.appendChild(button);
  button.setAttribute('type', 'submit');

  assert_false(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
  button.remove();
}, 'Setting same type via setAttribute while connected does not trigger UseCounters');

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  button.type = 'submit';

  assert_false(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
}, 'Setting same type via property while disconnected does not trigger UseCounters');

test(() => {
  resetUseCounters();
  const button = document.createElement('button');
  container.appendChild(button);
  button.type = 'submit';

  assert_false(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
  button.remove();
}, 'Setting same type via property while connected does not trigger UseCounters');

test(() => {
  resetUseCounters();
  container.innerHTML = '<button type="button">Click</button>';

  assert_false(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
  container.innerHTML = '';
}, 'Parser-set type does not trigger UseCounters');

test(() => {
  resetUseCounters();
  container.innerHTML = '<button type="button" id="btn">Click</button>';
  document.getElementById('btn').setAttribute('type', 'reset');

  assert_true(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
  container.innerHTML = '';
}, 'Changing parser-set type via setAttribute only triggers connected UseCounter');

test(() => {
  resetUseCounters();
  container.innerHTML = '<button type="button" id="btn">Click</button>';
  document.getElementById('btn').type = 'reset';

  assert_true(internals.isUseCounted(document, kTypeChangedConnected));
  assert_false(internals.isUseCounted(document, kTypeChangedDisconnected));
  container.innerHTML = '';
}, 'Changing parser-set type via property only triggers connected UseCounter');
</script>
